DocumentationResearch/requirements.txt
django~=5.0.6
requests~= 2.32.3
djangorestframework~=3.15.2
django_filter~=24.2
djoser~=2.2.3
psycopg2~=2.9.9
drf-spectacular[sidecar]
drf-yasg
